# app/docs/v1_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "fuge API",
        "title": "fuge API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/appointment": {
            "post": {
                "description": "创建预定时间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "创建预定时间",
                "parameters": [
                    {
                        "description": "预定日期",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models_v1.CreateBookingIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/appointment/{id}": {
            "get": {
                "description": "通过用户ID获取预定时间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "获取预定时间",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/room/store/{storeId}": {
            "get": {
                "description": "通过ID获取门店的房间列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1",
                    "门店相关"
                ],
                "summary": "获取门店的房间列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "门店ID",
                        "name": "storeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/login-wechat": {
            "post": {
                "description": "登陆微信",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "登陆微信",
                "parameters": [
                    {
                        "description": "登陆参数",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models_v1.LoginWeChatIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/work": {
            "post": {
                "description": "新建工作",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "新建工作",
                "parameters": [
                    {
                        "description": "创建工作内容参数",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models_v1.CreateServiceIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models_v1.CreateBookingIn": {
            "type": "object",
            "required": [
                "date",
                "endDateTime",
                "startDateTime"
            ],
            "properties": {
                "date": {
                    "type": "string",
                    "format": "date"
                },
                "endDateTime": {
                    "type": "string",
                    "format": "date-time"
                },
                "startDateTime": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "models_v1.CreateServiceIn": {
            "type": "object",
            "required": [
                "description",
                "name",
                "timePeriod"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "format": "string"
                },
                "name": {
                    "type": "string",
                    "format": "string"
                },
                "timePeriod": {
                    "type": "integer",
                    "format": "integer",
                    "minimum": 1,
                    "example": 60
                }
            }
        },
        "models_v1.LoginWeChatIn": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "format": "string"
                }
            }
        }
    }
}